move, focus and origin working
git-svn-id: https://tuxic.nl/laos/firmware/mbed/laos_server_menu@319 297e5b50-f6f4-49fd-be22-8ca01db2a49f

+++ LaosMenu/LaosMenu.h
@@ -72,8 +72,8 @@ private:
   unsigned char menu, ipfield, iofield;
   unsigned char powerfield, power[4];
   LaosDisplay *dsp;
-  float x,y,z;
-  float xoff, yoff, zoff;
+  int x,y,z;
+  int xoff, yoff, zoff;
   FILE *runfile;
   
 };
